[samza] Make a generic module for Venice nearline producers and remove tight coupling on Samza
