+++ internal/venice-nearline-producer/build.gradle
@@ -1,0 +1,23 @@
+configurations.implementation {
+    // jetty 6 conflicts with spark-java used in controller api
+    exclude group: 'org.mortbay.jetty'
+}
+
+dependencies {
+    // Schema related dependencies
+    implementation (project(':internal:venice-common')) {
+        //TODO, pull venice writer out of venice common so this is a smaller dependency
+        exclude module: 'kafka_2.10'
+        exclude group: 'org.scala-lang'
+    }
+
+    implementation project(':internal:venice-client-common')
+    implementation project(':clients:venice-thin-client')
+
+    implementation libraries.kafka
+    implementation libraries.log4j2api
+}
+
+ext {
+    jacocoCoverageThreshold = 0.01
+}
