+++ internal/venice-test-common/build.gradle
@@ -46,6 +46,7 @@ dependencies {
   implementation project(':services:venice-controller')
   implementation project(':services:venice-router')
   implementation project(':clients:venice-samza')
+  implementation project(':internal:venice-nearline-producer')
   implementation project(':internal:venice-client-common')
   implementation project(':services:venice-server')
   implementation project(':clients:venice-thin-client')
