+++ internal/venice-test-common/src/integrationtest/java/com/linkedin/venice/utils/IntegrationTestPushUtils.java
@@ -43,7 +43,7 @@ import com.linkedin.venice.pubsub.adapter.kafka.admin.ApacheKafkaAdminAdapterFac
 import com.linkedin.venice.pubsub.adapter.kafka.consumer.ApacheKafkaConsumerAdapterFactory;
 import com.linkedin.venice.pubsub.api.PubSubAdminAdapterFactory;
 import com.linkedin.venice.pubsub.api.PubSubConsumerAdapterFactory;
-import com.linkedin.venice.samza.VeniceObjectWithTimestamp;
+import com.linkedin.venice.producer.VeniceObjectWithTimestamp;
 import com.linkedin.venice.samza.VeniceSystemFactory;
 import java.util.Arrays;
 import java.util.Collections;
